+++ libs/vre/pages/search/advanced-search/jest.config.ts
@@ -4,4 +4,5 @@ export default {
   preset: '../../../../../jest.preset.js',
   coverageDirectory: '../../../../../coverage/libs/vre/pages/search/advanced-search',
   transformIgnorePatterns: ['node_modules/(?!@angular|@dasch-swiss|@ngrx|@ngx-translate)'],
+  collectCoverageFrom: ['src/**/*.ts', '!src/**/*.spec.ts', '!src/**/*.d.ts', '!src/test-setup.ts', '!src/index.ts'],
 };