+++ libs/vre/pages/search/search/src/lib/project-fulltext-search-page.component.ts
@@ -3,8 +3,8 @@ import { ComponentPortal } from '@angular/cdk/portal';
 import { AfterViewInit, Component, ElementRef, OnDestroy, ViewChild } from '@angular/core';
 import { FormBuilder } from '@angular/forms';
 import { ProjectPageService } from '@dasch-swiss/vre/pages/project/project';
+import { SearchTipsComponent } from '@dasch-swiss/vre/shared/app-common-to-move';
 import { map, startWith, Subject } from 'rxjs';
-import { SearchTipsComponent } from './search-tips.component';
 
 @Component({
   selector: 'app-project-fulltext-search-page',
@@ -33,22 +33,12 @@ import { SearchTipsComponent } from './search-tips.component';
 
     @if (query$ | async; as query) {
       <mat-divider />
-
-      <app-project-fulltext-search-result [query]="query" [projectId]="projectId" />
+      <div class="whole-height">
+        <app-search-result [query]="query" [projectId]="projectId" />
+      </div>
     }
   `,
-  styles: [
-    `
-      :host ::ng-deep .mat-mdc-form-field-subscript-wrapper {
-        display: none !important;
-      }
-
-      .big {
-        margin-top: 70px;
-        flex-direction: column;
-      }
-    `,
-  ],
+  styleUrls: ['./project-fulltext-search-page.component.scss'],
   standalone: false,
 })
 export class ProjectFulltextSearchPageComponent implements AfterViewInit, OnDestroy {