+++ libs/vre/pages/user-settings/user/src/lib/edit-password-dialog.component.ts
@@ -1,11 +1,22 @@
 import { Component, Inject, ViewChild } from '@angular/core';
-import { FormBuilder, FormControl, Validators } from '@angular/forms';
-import { MAT_DIALOG_DATA, MatDialogRef } from '@angular/material/dialog';
-import { MatStepper } from '@angular/material/stepper';
+import { FormBuilder, FormGroup, ReactiveFormsModule, Validators } from '@angular/forms';
+import { MatButton } from '@angular/material/button';
+import {
+  MAT_DIALOG_DATA,
+  MatDialogActions,
+  MatDialogClose,
+  MatDialogContent,
+  MatDialogRef,
+} from '@angular/material/dialog';
+import { MatStep, MatStepper } from '@angular/material/stepper';
 import { KnoraApiConnection, ReadUser } from '@dasch-swiss/dsp-js';
 import { UserApiService } from '@dasch-swiss/vre/3rd-party-services/api';
 import { DspApiConnectionToken } from '@dasch-swiss/vre/core/config';
 import { UserService } from '@dasch-swiss/vre/core/session';
+import { PasswordConfirmFormComponent, PasswordFormFieldComponent } from '@dasch-swiss/vre/shared/app-common-to-move';
+import { LoadingButtonDirective } from '@dasch-swiss/vre/ui/progress-indicator';
+import { DialogHeaderComponent } from '@dasch-swiss/vre/ui/ui';
+import { TranslateModule } from '@ngx-translate/core';
 import { finalize } from 'rxjs';
 
 export interface EditPasswordDialogProps {
@@ -37,7 +48,7 @@ export interface EditPasswordDialogProps {
         </mat-step>
 
         <mat-step [label]="'pages.userSettings.passwordForm.newPasswordStep' | translate">
-          <app-password-confirm-form (afterFormInit)="newPasswordControl = $event" />
+          <app-password-confirm-form (afterFormInit)="newPasswordFormGroup = $event" />
 
           <button
             mat-raised-button
@@ -63,13 +74,27 @@ export interface EditPasswordDialogProps {
       }
     `,
   ],
-  standalone: false,
+  standalone: true,
+  imports: [
+    DialogHeaderComponent,
+    TranslateModule,
+    MatDialogContent,
+    MatStepper,
+    MatStep,
+    PasswordFormFieldComponent,
+    MatButton,
+    PasswordConfirmFormComponent,
+    LoadingButtonDirective,
+    MatDialogActions,
+    MatDialogClose,
+    ReactiveFormsModule,
+  ],
 })
 export class EditPasswordDialogComponent {
   @ViewChild('stepper') stepper!: MatStepper;
 
   adminPasswordControl = this._fb.nonNullable.control('', [Validators.required]);
-  newPasswordControl!: FormControl<string>;
+  newPasswordFormGroup!: FormGroup;
 
   updateLoading = false;
   constructor(
@@ -96,16 +121,18 @@ export class EditPasswordDialogComponent {
   }
 
   updateNewPassword() {
-    this.newPasswordControl.markAllAsTouched();
+    this.newPasswordFormGroup.markAllAsTouched();
 
-    if (!this.newPasswordControl.valid) {
+    if (!this.newPasswordFormGroup.valid) {
       return;
     }
 
     this.updateLoading = true;
 
+    const newPassword = this.newPasswordFormGroup.get('password')?.value;
+
     this._userApiService
-      .updatePassword(this.data.user.id, this.adminPasswordControl.value, this.newPasswordControl.value)
+      .updatePassword(this.data.user.id, this.adminPasswordControl.value, newPassword)
       .pipe(
         finalize(() => {
           this.updateLoading = false;